user-function

+++ db/migrate/20200711040858_create_meals.rb
@@ -1,12 +1,12 @@
 class CreateMeals < ActiveRecord::Migration[5.2]
   def change
     create_table :meals do |t|
-      t.string :name
-      t.integer :protein
-      t.integer :fat
-      t.integer :carb
-      t.integer :cal
-      t.integer :user_id, foreign_key: true
+      t.string  :name
+      t.float :protein
+      t.float :fat
+      t.float :carb
+      t.float :cal
+      t.float :user_id, foreign_key: true
       t.timestamps
     end
   end
